Refactor startup/shutdown code
 * Unclutter commands package by moving stuff to startup module
 * Simplify/generalize commands code
 * Properly stop asyncio loop by cancelling all remaining tasks
 * Ignore CancelledError exceptions at shutdown
