+++ qtoggleserver/commands/shell.py
@@ -1,25 +1,20 @@
 #!/usr/bin/env python
 
-import asyncio
 import code
 
 from qtoggleserver import commands
 
 
-def execute() -> None:
+async def main() -> bool:
     import qtoggleserver  # noqa: F401; Required for locals()
 
-    loop = asyncio.get_event_loop()
-
-    loop.run_until_complete(commands.init())
-
-    commands.logger.info('starting interactive shell')
     code.interact(local=locals())
 
-    loop.run_until_complete(commands.cleanup())
-    loop.run_until_complete(loop.shutdown_asyncgens())
+    return False  # Don't run loop afterwards
 
-    commands.logger.info('bye!')
+
+def execute() -> None:
+    commands.execute(main())
 
 
 if __name__ == '__main__':
